Flattened scheduler exceptions into a single file
